+++ src/main/java/com/talearnt/Examples/ExampleController.java
@@ -2,7 +2,7 @@ package com.talearnt.Examples;
 
 
 import com.talearnt.Enums.ErrorCode;
-import com.talearnt.Utill.CommonResponse;
+import com.talearnt.Util.CommonResponse;
 import com.talearnt.swagger.annotions.UserNotFound;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
