+++ src/main/java/com/talearnt/util/CustomPhysicalNamingStrategy.java
@@ -1,0 +1,46 @@
+package com.talearnt.util;
+
+import org.hibernate.boot.model.naming.Identifier;
+import org.hibernate.boot.model.naming.PhysicalNamingStrategy;
+import org.hibernate.engine.jdbc.env.spi.JdbcEnvironment;
+
+public class CustomPhysicalNamingStrategy implements PhysicalNamingStrategy {
+
+    @Override
+    public Identifier toPhysicalCatalogName(Identifier name, JdbcEnvironment context) {
+        return apply(name);
+    }
+
+    @Override
+    public Identifier toPhysicalSchemaName(Identifier name, JdbcEnvironment context) {
+        return apply(name);
+    }
+
+    @Override
+    public Identifier toPhysicalTableName(Identifier name, JdbcEnvironment context) {
+        return apply(name);
+    }
+
+    @Override
+    public Identifier toPhysicalSequenceName(Identifier name, JdbcEnvironment context) {
+        return apply(name);
+    }
+
+    @Override
+    public Identifier toPhysicalColumnName(Identifier name, JdbcEnvironment context) {
+        return apply(name);
+    }
+
+    private Identifier apply(Identifier name) {
+        if (name == null || name.getText().isEmpty()) {
+            return name;
+        }
+
+        // 대문자 앞에 _ 추가, 그리고 전체 소문자로 변환
+        String newName = name.getText()
+                .replaceAll("([a-z])([A-Z])", "$1_$2")
+                .toLowerCase();
+
+        return Identifier.toIdentifier(newName);
+    }
+}
